refactor(keychainLib): basic IPasswordStore interface
no need for polyfills
allow for simple password store and disabling
