+++ src/lib/storage/index.ts
@@ -80,3 +80,45 @@ export interface IStorage {
   store: IStorageStore
   delete: IStorageDelete
 }
+
+export interface IPasswordStore {
+  savePassword: (password: string) => Promise<void>
+  getPassword: () => Promise<string>
+}
+
+
+/** TODO
+ * extract this into a separate npm module?
+ * it assumes node of course
+ */
+import { writeFileSync, readFileSync } from 'fs'
+
+export class FilePasswordStore implements IPasswordStore {
+  private filePath: string
+
+  constructor(filePath = './password.txt') {
+    this.filePath = filePath
+  }
+
+  public async getPassword() {
+    try {
+      return readFileSync(this.filePath).toString()
+    } catch (err) {
+      console.error('Error reading password file', err, '\n\n')
+      throw err
+    }
+  }
+
+  public async savePassword(pass: string) {
+    console.log('Saving password to: ', this.filePath)
+    return writeFileSync(this.filePath, pass)
+  }
+}
+
+export class EmptyPasswordStore implements IPasswordStore {
+  public async getPassword() {
+    return Buffer.from('aaaaaaaaaaaaaaaa', 'hex').toString('base64')
+  }
+  public async savePassword() {
+  }
+}
